Move Azure API Client back to init
